requirements.txt
streamlit
tensorflow
numpy
pandas
pillow
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
scikit-learn
